--- src/core/data-parse-v3.ts
@@ -1,12 +1,0 @@
-import { OpenAPIV3 } from 'openapi-types'
-
-import { SwaggerJsonUrlItem } from '../tools'
-
-export function parseSwaggerJsonV3(
-  swaggerJson: OpenAPIV3.Document,
-  configItem: SwaggerJsonUrlItem
-): SwaggerJsonTreeItem[] {
-  // console.log(swaggerJson, configItem)
-
-  return []
-}
