+++ src/core/index.ts
@@ -1,4 +1,5 @@
 export * from './base'
 export * from './data-fetch'
-export * from './data-parse'
-export * from './data-parse-v3'
+export * from './data-parser'
+export * from './data-parser-v3'
+export * from './data-parser-v2'
